+++ web/ts/data-store/data-store.ts
@@ -1,0 +1,5 @@
+import {VideoSectionProvider} from "./video-sections";
+
+export abstract class DataStore {
+    static videoSections : VideoSectionProvider = new VideoSectionProvider();
+}
